+++ examples/geometry/volumes/CMakeLists.txt
@@ -2,7 +2,7 @@ add_subdirectory(distance)
 
 set( DGTAL_EXAMPLES_SRC
         dvcm-2d
-        fullConvexitySimplicity2D
+        fullConvexityCollapsiblePoints2D
         fullConvexityLUT2D
 )
